requirements.txt
matplotlib==3.7.3
moviepy==1.0.3
numpy==1.24.4
opencv_python_headless==4.9.0.80
Pillow==10.3.0
pycocotools==2.0.7
PyQt5==5.15.10
PyQt5_sip==12.13.0
pytorch_msssim==1.0.0
scipy==1.14.0
tqdm==4.66.1
